# openapi.yaml
...
openapi: 3.0.3
info:
    title: App API
    version: 0.0.1
paths:
    /api/admin_dhb/deposit:
        get:
            tags:
                - App
            operationId: App_Deposit
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DepositReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/admin_dhb/fee:
        get:
            tags:
                - App
            operationId: App_AdminFee
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AdminFeeReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/admin_dhb/withdraw:
        get:
            tags:
                - App
            description: |-
                rpc AdminRewardList (AdminRewardListRequest) returns (AdminRewardListReply) {
                		option (google.api.http) = {
                			get: "/api/admin_dhb/reward_list"
                		};
                	};

                	rpc AdminUserList (AdminUserListRequest) returns (AdminUserListReply) {
                		option (google.api.http) = {
                			get: "/api/admin_dhb/user_list"
                		};
                	};

                	rpc AdminLocationList (AdminLocationListRequest) returns (AdminLocationListReply) {
                		option (google.api.http) = {
                			get: "/api/admin_dhb/location_list"
                		};
                	};

                	rpc AdminWithdrawList (AdminWithdrawListRequest) returns (AdminWithdrawListReply) {
                		option (google.api.http) = {
                			get: "/api/admin_dhb/withdraw_list"
                		};
                	};
            operationId: App_AdminWithdraw
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AdminWithdrawReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/admin_dhb/withdraw_eth:
        get:
            tags:
                - App
            operationId: App_AdminWithdrawEth
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AdminWithdrawEthReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/app_server/delete_balance_reward:
        post:
            tags:
                - App
            operationId: App_DeleteBalanceReward
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/DeleteBalanceRewardRequest_SendBody'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeleteBalanceRewardReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/app_server/eth_authorize:
        post:
            tags:
                - App
            operationId: App_EthAuthorize
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/EthAuthorizeRequest_SendBody'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/EthAuthorizeReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/app_server/fee_reward_list:
        get:
            tags:
                - App
            operationId: App_FeeRewardList
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/FeeRewardListReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/app_server/recommend_list:
        get:
            tags:
                - App
            operationId: App_RecommendList
            parameters:
                - name: type
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/RecommendListReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/app_server/recommend_reward_list:
        get:
            tags:
                - App
            operationId: App_RecommendRewardList
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/RecommendRewardListReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/app_server/recommend_update:
        post:
            tags:
                - App
            operationId: App_RecommendUpdate
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/RecommendUpdateRequest_SendBody'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/RecommendUpdateReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/app_server/reward_list:
        get:
            tags:
                - App
            operationId: App_RewardList
            parameters:
                - name: type
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/RewardListReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/app_server/set_balance_reward:
        post:
            tags:
                - App
            operationId: App_SetBalanceReward
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SetBalanceRewardRequest_SendBody'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SetBalanceRewardReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/app_server/user_info:
        get:
            tags:
                - App
            operationId: App_UserInfo
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UserInfoReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/app_server/withdraw:
        post:
            tags:
                - App
            operationId: App_Withdraw
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/WithdrawRequest_SendBody'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/WithdrawReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
components:
    schemas:
        AdminFeeReply:
            type: object
            properties: {}
        AdminWithdrawEthReply:
            type: object
            properties: {}
        AdminWithdrawReply:
            type: object
            properties: {}
        DeleteBalanceRewardReply:
            type: object
            properties:
                status:
                    type: string
        DeleteBalanceRewardRequest_SendBody:
            type: object
            properties:
                amount:
                    type: string
        DepositReply:
            type: object
            properties: {}
        EthAuthorizeReply:
            type: object
            properties:
                token:
                    type: string
...
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...